requirements.txt
Django==3.1.4
django-extensions==3.1.0
djangorestframework==3.12.2
ulid-py==1.1.0
redis==3.5.3
